+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/activity/ActivityContext.kt
@@ -1,6 +1,7 @@
 package io.github.pandier.intellijdiscordrp.activity
 
 import com.intellij.openapi.application.ApplicationInfo
+import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.ApplicationNamesInfo
 import com.intellij.openapi.application.ReadAction
 import com.intellij.openapi.editor.Editor
@@ -8,6 +9,7 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ProjectFileIndex
 import com.intellij.openapi.vfs.VfsUtil
 import com.intellij.openapi.vfs.VirtualFile
+import git4idea.GitUtil
 import io.github.pandier.intellijdiscordrp.service.TimeTrackingService
 import io.github.pandier.intellijdiscordrp.settings.DiscordSettings
 import io.github.pandier.intellijdiscordrp.settings.discordSettingsComponent
@@ -26,33 +28,42 @@ class ActivityFileContext(
     val typeName: String,
     val line: Int?,
     val lineCount: Int?,
+    val length: Long?,
+    val start: Instant,
 )
 
 class ActivityContext(
-    val start: Instant,
     val appName: String,
     val appFullName: String,
     val appVersion: String,
+    val appStart: Instant,
     val project: WeakReference<Project>,
     val projectName: String,
+    val projectRepositoryUrl: String?,
+    val projectStart: Instant,
     val file: ActivityFileContext? = null,
 ) {
     companion object Factory {
         fun create(
             project: Project,
             file: VirtualFile? = null,
-            editor: Editor? = null,
-            start: Instant = TimeTrackingService.getInstance(project).start,
+            editor: Editor? = null
         ): ActivityContext {
+            val timeTrackingService = TimeTrackingService.getInstance()
             val appInfo = ApplicationInfo.getInstance()
             val appNames = ApplicationNamesInfo.getInstance()
+            val repositoryManager = GitUtil.getRepositoryManager(project)
+            val repository = file?.let { repositoryManager.getRepositoryForFileQuick(it) }
+                ?: repositoryManager.repositories.firstOrNull()
             return ActivityContext(
-                start = start,
                 appName = appNames.fullProductName,
                 appFullName = appNames.fullProductNameWithEdition,
                 appVersion = appInfo.fullVersion,
-                projectName = project.name,
+                appStart = timeTrackingService.getOrInit(ApplicationManager.getApplication()),
                 project = WeakReference(project),
+                projectName = project.name,
+                projectRepositoryUrl = repository?.remotes?.let(GitUtil::getDefaultOrFirstRemote)?.firstUrl,
+                projectStart = timeTrackingService.getOrInit(project),
                 file = file?.let {
                     val contentRoot = ReadAction.compute<VirtualFile?, Exception> {
                         ProjectFileIndex.getInstance(project).getContentRootForFile(it)
@@ -67,6 +78,8 @@ class ActivityContext(
                         typeName = activityFileType?.friendlyName ?: it.fileType.name,
                         line = editor?.caretModel?.logicalPosition?.line?.plus(1),
                         lineCount = editor?.document?.lineCount?.let { max(it, 1) },
+                        length = it.length,
+                        start = timeTrackingService.getOrInit(it),
                     )
                 }
             )
@@ -76,13 +89,13 @@ class ActivityContext(
     /**
      * Renders this activity context to an [Activity] using global plugin settings
      * and plugin settings of the project with respect to settings hiding the activity.
-     * Returns null if the reference to the projet was already dropped.
+     * Returns null if the reference to the project was already dropped.
      */
     fun createActivity(): Activity? {
-        return createActivity(
-            discordSettingsComponent.state,
-            project.get()?.discordProjectSettingsComponent?.state
-        )
+        val projectSettings = project.get()?.discordProjectSettingsComponent?.state
+        if (projectSettings == null)
+            return null
+        return createActivity(discordSettingsComponent.state, projectSettings)
     }
 
     /**
@@ -93,14 +106,14 @@ class ActivityContext(
      */
     fun createActivity(
         settings: DiscordSettings,
-        projectSettings: DiscordProjectSettings?,
+        projectSettings: DiscordProjectSettings,
     ): Activity? {
-        if (projectSettings == null || !projectSettings.showRichPresence)
+        if (!projectSettings.showRichPresence)
             return null
         val displayMode = ActivityDisplayMode.getSupportedFrom(
             projectSettings.displayMode ?: settings.defaultDisplayMode,
             this
         )
-        return settings.getActivityFactory(displayMode).create(this)
+        return settings.createActivityFactory(displayMode, projectSettings).create(this)
     }
 }