+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/settings/project/DiscordProjectSettings.kt
@@ -4,5 +4,6 @@ import io.github.pandier.intellijdiscordrp.activity.ActivityDisplayMode
 
 data class DiscordProjectSettings(
     var showRichPresence: Boolean = true,
+    var showRepoButton: Boolean = true,
     var displayMode: ActivityDisplayMode? = null,
 )